requirements.txt
# phytoDB - Analyse PPP Côte-d'Or
# Python 3.10+

requests>=2.28.0
pyyaml>=6.0
